+++ jni/sandbox/svs/src/svs_constants.cpp
@@ -1,0 +1,16 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include "svs_constants.h"
+
+const std::string knn_jni::CONSTRUCTION_WINDOW_SIZE = "construction_window_size";
+const std::string knn_jni::ALPHA = "alpha";
+
+const std::string knn_jni::SEARCH_WINDOW_SIZE = "search_window_size";
+const std::string knn_jni::SEARCH_BUFFER_CAPACITY = "search_buffer_capacity";
+
+const std::string knn_jni::ENCODER = "encoder";
+const std::string knn_jni::LEANVEC_TRAINING_THRESHOLD = "training_threshold";
+const std::string knn_jni::LEANVEC_ROUGH_TRAINING_THRESHOLD = "rough_training_threshold";